components restructure/introduce isOpen
