+++ packages/plugins/io-center/src/plugin.svelte
@@ -36,12 +36,15 @@ import ObjectTree from './ui/components/object-tree/object-tree.svelte';
 const objectTree = [
     {
       name: 'Device1',
+	  isOpen: true,
       children: [
         {
           name: 'Node1',
+		  isOpen: true,
           children: [
             {
               name: 'ObjectInstance1',
+			  isOpen: true,
               children: [
                 { name: 'Attribute1' },
                 { name: 'Attribute2' },
@@ -49,6 +52,7 @@ const objectTree = [
             },
             {
               name: 'ObjectInstance2',
+			  isOpen: false,
               children: [
                 { name: 'Attribute3' },
                 { name: 'Attribute4' },
@@ -58,9 +62,11 @@ const objectTree = [
         },
         {
           name: 'Node2',
+		  isOpen: false,
           children: [
             {
               name: 'ObjectInstance3',
+			  isOpen: false,
               children: [
                 { name: 'Attribute5' },
                 { name: 'Attribute6' },
@@ -72,12 +78,15 @@ const objectTree = [
     },
     {
       name: 'Device2',
+	  isOpen: true,
       children: [
         {
           name: 'Node3',
+		  isOpen: true,
           children: [
             {
               name: 'ObjectInstance4',
+	  		  isOpen: true,
               children: [
                 { name: 'Attribute7' },
                 { name: 'Attribute8' },
@@ -87,9 +96,11 @@ const objectTree = [
         },
         {
           name: 'Node4',
+	  	  isOpen: true,
           children: [
             {
               name: 'ObjectInstance5',
+			  isOpen: true,
               children: [
                 { name: 'Attribute9' },
                 { name: 'Attribute10' },
