+++ packages/plugins/io-center/src/ui/components/object-tree/object-tree.svelte
@@ -1,16 +1,28 @@
 <script lang="ts">
-	import type { ObjectTree, TreeNode } from "./object-tree";
-	import TreeItem from "./TreeItem.svelte";
+	import type { ObjectTree } from "./object-tree";
+	import { Accordion } from "@oscd-plugins/core-ui-svelte";
+	import type { TreeNode as TreeNodeType } from "./object-tree";
+	import TreeNode from "./tree-node.svelte";
 
 	type Props = {
 		objectTree: ObjectTree;
 	};
 
 	let { objectTree }: Props = $props();
+
+	const getOpenItems = (nodes: TreeNodeType[]): string[] =>
+		nodes.reduce<string[]>(
+			(acc, node) => [
+				...acc,
+				...(node.isOpen ? [node.name] : []),
+				...(node.children ? getOpenItems(node.children) : []),
+			],
+			[],
+		);
 </script>
 
-<div class="w-1/4">
-	{#each objectTree as node}
-		<TreeItem treeNode={node} />
+<Accordion.Root value={getOpenItems(objectTree)} type="multiple" class="w-1/4">
+	{#each objectTree as treeNode}
+		<TreeNode {treeNode} />
 	{/each}
-</div>
+</Accordion.Root>
