+++ packages/plugins/io-center/src/ui/components/object-tree/object-tree.ts
@@ -3,6 +3,7 @@ export type ObjectTree = TreeNode[]
 
 export type TreeNode = {
 	name: string
+	isOpen?: boolean
 	children?: TreeNode[]
 }
